requirements.txt
# Core ML
torch>=2.2.0
transformers>=4.40.0
peft>=0.10.0
bitsandbytes>=0.43.0
accelerate>=0.28.0
datasets>=2.18.0
trl>=0.8.6

# API
fastapi==0.104.0
uvicorn==0.24.0
pydantic==2.5.0

# AWS
boto3==1.34.0
sagemaker==2.199.0

# Data & ML
pyyaml==6.0.1
python-dotenv==1.0.0
sentence-transformers==2.2.2
scikit-learn==1.3.0
scipy==1.11.0
numpy==1.24.0
pandas==2.0.0

# GitHub API
PyGithub==2.1.0

# Utils
tqdm==4.66.0
rich==13.7.0
requests==2.31.0

# Development
pytest==7.4.0
black==23.0.0
flake8==6.0.0